+++ Src/Server/MK.Booking/MK.Booking.Commands/RedeemPromotion.cs
@@ -12,7 +12,7 @@ namespace apcurium.MK.Booking.Commands
 
         public Guid PromoId { get; set; }
         public Guid OrderId { get; set; }
-        public double TotalAmountOfOrder { get; set; }
+        public decimal TotalAmountOfOrder { get; set; }
 
         public Guid Id { get; private set; }
     }
