+++ Src/Server/MK.Booking/MK.Booking.Events/PromotionUpdated.cs
@@ -24,7 +24,7 @@ namespace apcurium.MK.Booking.Events
 
         public bool AppliesToFutureBooking { get; set; }
 
-        public double DiscountValue { get; set; }
+        public decimal DiscountValue { get; set; }
 
         public PromoDiscountType DiscountType { get; set; }
